+++ CHANGELOG.md
@@ -19,6 +19,8 @@
   well as data files)
 - Select controls can be disabled now
 - Bump CodeMirror dependencies
+- Define a new shared type, `IPCAPI` that can be used to type the various IPC
+  APIs the service providers use across the app.
 
 # 3.3.1